Adding append and prepend to html() function
Although the after and before function variables worked fine, I wanted to append an object into the element instead of before or after it.. so this is what I came up with. Its a simple update, but could be useful for other people to.

I didnt create the minified file so you may want to re-compile that if this gets approved.

By the way, thanks for starting this! Its short and sweet to the point. Just what I was looking for :)